+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.0
+
+* Raised Target SDK Version to `33` (Android 13 | Codename: `TIRAMISU`)
+* Added access to `canPackageQuery` method which is exclusive to Android 13
+* Added support for non-deprecated alternatives of some APIs on Android 13
+* Raised plugin's Gradle version to `7.2.2`
+
 ## 0.2.0
 
 * Fixed `getInstalledPackages` results having null `applicationInfo` (reported issue [here](https://github.com/nsNeruno/android_package_manager/issues/1) | credits to [@pablo8899](https://github.com/pablo8899))